+++ Changelog.md
@@ -1,6 +1,7 @@
 * XXXXX: v1.8:
   * `findso`: better help, fixes. `-s` flag to skip symlinks.
   * `tsdec`: add Chome/WebKit timestamp support.
+  * `symlink` command to create symlinks for all applets
 * 2024-11-11: v1.7:
   * new `escape` and `unescape` applets for easy string (un)escaping
   * `base`: default to hex output for decimal input